ThemeEditor: Show unnecessary scrollbars :(
There's some bug that shows the scrollbars, which from what I think it's
only because the available space (which is the full widget here) *with*
the scrollbars included take more size than the widget space.

It stays that way until the second draw/layout/whatever, but since
ThemeEditor has resizing disabled it always stays that way.